test(input): migrate legacy clear-on-edit
